+++ tests/errors/mod.rs
@@ -75,12 +75,9 @@ async fn test_default_404_with_redirect() {
     let resp = resp_result.unwrap();
     assert_eq!(
         resp.status(),
-        http::StatusCode::MOVED_PERMANENTLY,
-        "/i-do-not-exist should return 301"
+        http::StatusCode::NOT_FOUND,
+        "/i-do-not-exist should return 404"
     );
-
-    let location = resp.headers().get(http::header::LOCATION).unwrap();
-    assert_eq!(location, "/i-do-not-exist/");
 
     let resp_result = req_path("/i-do-not-exist/").await;
     let resp = resp_result.unwrap();